@@ -9,6 +9,10 @@
     "picturefill": "3.0.2",
     "waypoints": "4.0.1"
   },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack"
+  },
   "devDependencies": {
     "autoprefixer": "7.1.6",
     "babel-core": "6.26.0",
@@ -20,7 +24,7 @@
     "gulp-cssnano": "2.1.2",
     "gulp-imagemin": "4.0.0",
     "gulp-modernizr": "1.0.0-alpha",
-    "gulp-postcss": "7.0.0",
+    "gulp-postcss": "^7.0.0",
     "gulp-rename": "1.2.2",
     "gulp-rev": "8.1.0",
     "gulp-svg-sprite": "1.3.7",
@@ -31,8 +35,10 @@
     "postcss-hexrgba": "1.0.0",
     "postcss-import": "11.0.0",
     "postcss-mixins": "6.2.0",
-    "postcss-nested": "2.1.2",
+    "postcss-nested": "^2.1.2",
     "postcss-simple-vars": "4.1.0",
-    "webpack": "3.8.1"
+    "webpack": "^3.12.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-command": "^0.4.1"
   }
 }
