@@ -1,1 +1,3 @@
-var Person = require('./modules/Person');
+import MobileMenu from './modules/MobileMenu';
+
+var mobileMenu = new MobileMenu();
